% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assortativity.R
\name{dw_assort}
\alias{dw_assort}
\title{Compute the assortativity coefficient of a weighted and directed network.}
\usage{
dw_assort(adj, type = c("out-in", "in-in", "out-out", "in-out"))
}
\arguments{
\item{adj}{is an adjacency matrix of an weighted and directed network.}

\item{type}{which type of assortativity coefficient to compute: "out-in" (default), 
"in-in", "out-out" or "in-out"?}
}
\value{
a scalar of assortativity coefficient
}
\description{
Compute the assortativity coefficient of a weighted and directed network.
}
\note{
When the adjacency matrix is binary (i.e., directed but unweighted networks), \code{dw_assort}
returns the assortativity coefficient proposed in Foster et al. (2010).
}
\examples{
## Generate a network according to the Erd\"{o}s-Renyi model of order 20
## and parameter p = 0.3
edge_ER <- rbinom(400,1,0.3)
weight_ER <- sapply(edge_ER, function(x) x*sample(3,1))
adj_ER <- matrix(weight_ER,20,20)
system.time(myassort <- dw_assort(adj_ER, type = "out-in"))
myassort

}
\references{
\itemize{
\item Foster, J.G., Foster, D.V., Grassberger, P. and Paczuski, M. (2010). Edge direction 
and the structure of networks. \emph{Proceedings of the National Academy of Sciences of the
United States}, 107(24), 10815--10820.
\item Yuan, Y. Zhang, P. and Yan, J. (2020+). Assortativity coefficients for 
weighted and directed networks
}
}
